Add function to calculate profit if seats are sold out
Depending on the number of rows and columns, an algorithm
is used to calculate the total profit. If the number of seats is less than or equal to 60, each seats provides an income of $10. If the number is greater, the front half (rounded down) of the seats provide $10 each while the remating back seats provide $8 each.