Add functions to clean up the main function
A function that performs input validation on integer values was created to eliminate repetition when it came to user input validation. The function not only checks if an integer value was entered but also checks to see if it is in the necessary range.

Multiple functions were created such that the start of the user interface (menu) would not be controlled by the main function. Instead, it was integrated into the Theater class.