Add ticket sales statistics and increase readability
A statistics function was added to print out ticket sales as well as income. This would allow cinema managers to have a way to visualize their profits.

Additionally, print statements were formatted to increase readability of the console output allowing for users to have an easier time navigating the program.