Update customer interface such that theater can be chosen from list
When the user decides to purchase a ticket, the user is taken to a list of available theater rooms (still need to add showing movies). The user then enters a theater room number which takes them to that theater's specific seating chart where they can choose their seats. Entering 0 will take them back to the menu.